requirements.txt
aiohttp==3.7.4.post0	
aiosignal==1.2.0	
amqp==5.1.0	
asgiref==3.4.1	
async-timeout==3.0.1	
attrs==21.2.0	
backports.zoneinfo==0.2.1	
base58==2.1.1	
billiard==3.6.4.0	
bitarray==1.2.2	
boto3==1.20.23	
botocore==1.23.23	
bytecode==0.13.0	
celery==5.2.6	
certifi==2021.10.8	
cffi==1.15.1	
chardet==4.0.0	
charset-normalizer==2.0.7	
click==8.1.2	
click-didyoumean==0.3.0	
click-plugins==1.1.1	
click-repl==0.2.0	
cryptography==38.0.1	
cytoolz==0.11.2	
ddsketch==2.0.4	
ddtrace==1.4.4	
Deprecated==1.2.13	
discord.py==1.7.3	
dj-database-url==0.5.0	
Django==4.0.1	
django-bmemcached==0.3.0	
django-celery-beat @ git+https://github.com/celery/django-celery-beat@30185a943ed2d4624c05ea25404eef6204829953	
django-cors-headers==3.10.0	
django-debug-toolbar==3.2.4	
django-filter==21.1	
django-heroku==0.3.1	
django-ratelimit==3.0.1	
django-storages==1.12.3	
django-timezone-field==4.2.3	
djangorestframework==3.12.4	
djangorestframework-api-key==2.1.0	
eth-abi==2.1.1	
eth-account==0.5.6	
eth-hash==0.3.2	
eth-keyfile==0.5.1	
eth-keys==0.3.3	
eth-rlp==0.2.1	
eth-typing==2.2.2	
eth-utils==1.10.0	
frozenlist==1.2.0	
greenlet==1.1.2	
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0	
hexbytes==0.2.2	
idna==3.3	
image==1.5.33	
ipfs-api==0.2.3	
ipfshttpclient==0.8.0a2	
jmespath==0.10.0	
jsonschema==3.2.0	
kombu==5.2.4	
lru-dict==1.1.7	
Markdown==3.3.4	
multiaddr==0.0.9	
multidict==5.2.0	
netaddr==0.8.0	
numpy==1.22.3	
packaging==21.3	
parsimonious==0.8.1	
Pillow==8.4.0	
prompt-toolkit==3.0.29	
protobuf==3.20.1	
psutil==5.9.2	
psycopg2==2.9.2	
py-solc-x==1.1.1	
pycparser==2.21	
pycryptodome==3.11.0	
pylibmc==1.5.2	
pymemcache==3.5.2	
pyOpenSSL==22.0.0	
pyparsing==3.0.7	
pyrsistent==0.18.0	
python-binary-memcached==0.31.1	
python-crontab==2.6.0	
python-dateutil==2.8.2	
python-http-client==3.3.4	
python-magic==0.4.25	
pytz==2021.3	
PyYAML==6.0	
redis==4.1.4	
requests==2.26.0	
rlp==2.0.1	
s3transfer==0.5.0	
scout-apm==2.24.2	
semantic-version==2.9.0	
sendgrid==6.9.3	
six==1.16.0	
sqlparse==0.4.2	
starkbank-ecdsa==2.0.3	
tenacity==8.0.1	
toolz==0.11.2	
typing_extensions==4.0.0	
uhashring==2.1	
urllib3==1.26.7	
varint==1.0.2	
vine==5.0.0	
wcwidth==0.2.5	
web3==5.24.0	
websockets==9.1	
whitenoise==5.3.0	
wrapt==1.14.0	
yarl==1.7.2	
zope.event==4.5.0	
zope.interface==5.4.0
